requirements.txt
python-multipart
jinja2
uvicorn
bcrypt
fastapi
PyJwt
python-jose
passlib
pydantic
pymongo
